+++ clients/admin-ui/src/features/datamap/reporting/datamap-report-context.tsx
@@ -9,9 +9,8 @@ import {
 import { useLocalStorage } from "~/features/common/hooks/useLocalStorage";
 import { DATAMAP_GROUPING } from "~/types/api";
 
-import { DATAMAP_LOCAL_STORAGE_KEYS } from "../constants";
 import { DatamapReportFilterSelections } from "../types";
-import { COLUMN_IDS } from "./DatamapReportTableColumns";
+import { COLUMN_IDS, DATAMAP_LOCAL_STORAGE_KEYS } from "./constants";
 import { getColumnOrder } from "./utils";
 
 interface DatamapReportContextProps {
@@ -27,6 +26,8 @@ interface DatamapReportContextProps {
   setColumnVisibility: Dispatch<SetStateAction<Record<string, boolean>>>;
   columnSizing: Record<string, number>;
   setColumnSizing: Dispatch<SetStateAction<Record<string, number>>>;
+  columnNameMapOverrides: Record<string, string>;
+  setColumnNameMapOverrides: Dispatch<SetStateAction<Record<string, string>>>;
 }
 
 export const DatamapReportContext = createContext<DatamapReportContextProps>(
@@ -42,10 +43,12 @@ export const DatamapReportProvider = ({
     DATAMAP_LOCAL_STORAGE_KEYS.CUSTOM_REPORT_ID,
     "",
   );
+
   const [groupBy, setGroupBy] = useLocalStorage<DATAMAP_GROUPING>(
     DATAMAP_LOCAL_STORAGE_KEYS.GROUP_BY,
     DATAMAP_GROUPING.SYSTEM_DATA_USE,
   );
+
   const [selectedFilters, setSelectedFilters] =
     useLocalStorage<DatamapReportFilterSelections>(
       DATAMAP_LOCAL_STORAGE_KEYS.FILTERS,
@@ -55,19 +58,27 @@ export const DatamapReportProvider = ({
         dataCategories: [],
       },
     );
+
   const [columnOrder, setColumnOrder] = useLocalStorage<string[]>(
     DATAMAP_LOCAL_STORAGE_KEYS.COLUMN_ORDER,
     getColumnOrder(groupBy),
   );
+
   const [columnVisibility, setColumnVisibility] = useLocalStorage<
     Record<string, boolean>
   >(DATAMAP_LOCAL_STORAGE_KEYS.COLUMN_VISIBILITY, {
     [COLUMN_IDS.SYSTEM_UNDECLARED_DATA_CATEGORIES]: false,
     [COLUMN_IDS.DATA_USE_UNDECLARED_DATA_CATEGORIES]: false,
   });
+
   const [columnSizing, setColumnSizing] = useLocalStorage<
     Record<string, number>
   >(DATAMAP_LOCAL_STORAGE_KEYS.COLUMN_SIZING, {});
+
+  const [columnNameMapOverrides, setColumnNameMapOverrides] = useLocalStorage<
+    Record<string, string>
+  >(DATAMAP_LOCAL_STORAGE_KEYS.COLUMN_NAMES, {});
+
   const contextValue: DatamapReportContextProps = useMemo(
     () => ({
       savedCustomReportId,
@@ -82,6 +93,8 @@ export const DatamapReportProvider = ({
       setColumnVisibility,
       columnSizing,
       setColumnSizing,
+      columnNameMapOverrides,
+      setColumnNameMapOverrides,
     }),
     [
       savedCustomReportId,
@@ -96,6 +109,8 @@ export const DatamapReportProvider = ({
       setColumnVisibility,
       columnSizing,
       setColumnSizing,
+      columnNameMapOverrides,
+      setColumnNameMapOverrides,
     ],
   );